TACHYON-198 Move store/load to Hadoop conf methods to a utils class per Haoyuan suggestion.
